// OneDrive/Desktop/Projects/typescript-task17/src/Components/types.ts
export interface IHuman {
  gender: string;
  DOB: string;
  status: boolean;
}

export interface IPerson {
  name: String;
  age: number | string;
}

export interface INumber {
 dates: number[];
}

export interface ICar {
  make: string;
  model: string;
}

export interface IBike {
  brand: string;
  type: string;
}